[REFACTOR] Move getSunstoneStage to a leaf to break the Sunstone import cycle
Sunstone.tsx imports its Recovered/Depleting/Depleted children, and the children
imported getSunstoneStage back from Sunstone.tsx. Extracting it into its own leaf
module (sunstone/getSunstoneStage.ts) breaks that cycle — mirrors the crimstone fix.
